StringFigures.jl

Documentation for StringFigures.jl

Linear Sequences and knot diagrams

StringFigures.FrameNode — Type

A FrameNode represents both:

A loop on a finger or other body part that is holding the string.

A "puncture" of the plane that the string cannot freely cross.

It is parametrized by nodetype ∈ (:L, :R), index::Int, loop::Int

StringFigures.CrossNode — Type

A CrossNode represents a crossing in a knot diagram, and is parametrized by index::Int and nodetype ∈ (:U,:O)

StringFigures.LinearSequence — Type

A LinearSequence represents a knot in punctured space with a notation similar to Gauss codes. A linear sequence is a sequence (separated by :) of either FrameNodes (loosely, fingers) and CrossNodes. CrossNodes appear in pairs in the sequence (one with the attribute :O (for "over") and one with the attribute :U (for "under")

See also: seq"", plot, FrameNode, CrossNode

StringFigures.@node_str — Macro

node"xxx" generates either a FrameNode or a CrossNode.

julia> node"L1"
node"L1"

julia> show(node"L1")
L1

julia> node"x100(0)"
node"x100(0)"

julia> show(node"x100(U)")
x100(U)

StringFigures.@seq_str — Macro

seq"xxx" returns the LinearSequence "xxx".

See also: open"", storer""

Examples

julia> seq"L1:R1:R2"
seq"L1:R1:R2"

StringFigures.@open_str — Macro

open"xxx" returns the opening xxx.

Examples

julia> open"O0"
seq"L2:R2"

julia> open"O1"
seq"L1:L5:R5:R1"

julia> open"OA"
seq"L1:x1(0):R2:x2(0):L5:R5:x2(U):L2:x1(U):R1"

StringFigures.@storer_str — Macro

storer"xxx" is equivalent to seq"xxx" but allows some fuzziness to be able to easily copy-paste from Storer's OCR'd book :)

Examples

julia> osage3diamondsfig41 = storer"Ll: x1(0): x2(U): x3(0): x4(0): x5(U): x6(0): x7(0): x8(UJ: x9(U):
       x10(0): x11(0): x12(U): x13(UJ: x14(0): x15(U): x16(U): R2: x16(0):
       x17(U): x18{0): x19(U): x20(0): x5(0): x4(U): x21(0): L2: x2l(U):
       x3(U): x2(0): xl(U): x22(U): x23(0): x6(U): x20(U): x19(0): x9(0):
       xlO (U): x18 (U): xl7(0): x15 (0): x14 (U): x13 (0): Rl: x12'(0): xll (U):
       x8(0): x7(UJ: x23(U): x22(0) "
seq"L1:x1(0):x2(U):x3(0):x4(0):x5(U):x6(0):x7(0):x8(U):x9(U):x10(0):x11(0):x12(U):x13(U):x14(0):x15(U):x16(U):R2:x16(0):x17(U):x18(0):x19(U):x20(0):x5(0):x4(U):x21(0):L2:x21(U):x3(U):x2(0):x1(U):x22(U):x23(0):x6(U):x20(U):x19(0):x9(0):x10(U):x18(U):x17(0):x15(0):x14(U):x13(0):R1:x12(0):x11(U):x8(0):x7(U):x23(U):x22(0)"

StringFigures.plot — Function

plot(p::LinearSequence; rfact, k, randomize, labels, shadowc, kwd...)

Plots p using Tutte embedding. Parallel edges are then separated by slightly translating them perpendicularly to the segment joining the two vertices.

fact::Float64: Multiplicative factor for edge width

rfact::Float64: Distance between parallel edges (0.02)

randomize::Bool: Slightly randomize positions (false)

labels::Bool: Add labels to the plot (true)

crossings::Bool: Should crossings be plotted

shadowc::colorant: Color of the string shadow (colorant"black")

stringc::colorant: Color of the string (colorant"white")

kwd...: Additional options for gplot ((;)). See e.g. background_color.

Example1: plot(proc"OA::DL1|") Example2: plot(proc"OA::DL1|"; shadowc="white", stringc="black", nodelabelc="black", background_color="white")

StringFigures.release — Function

release(p::LinearSequence, f::FrameNode)

Releases f from the LinearSequence p

StringFigures.twist — Function

twist(p::LinearSequence, f::FrameNode, away::Bool)

Twists the top loop of f in LinearSequence p, away from the executer.

StringFigures.pick — Function

pick(p::LinearSequence, over::Bool, away::Bool, f::FrameNode, arg::FrameNode, near::Bool, above::Bool=false)

Picks with finger f, over intermediate strings, near string in the loop arg, from above.

StringFigures.simplify — Function

Extension-cancellation simplifications

Passages

StringFigures.Passage — Type

The Passage type represents one passage or move in a string figure construction. A specific Passage can be applied to a LinearSequence as a function.

See also: ReleasePassage, TwistPassage, PickPassage, ExtendPassage, MultiPickPassage, NavahoPassage, PowerPassage, pass""

StringFigures.ReleasePassage — Type

A ReleasePassage represents the release of one loop. It is denoted by the "□" symbol in Storer, which we represent in ASCII with "D" (for delete)

julia> pass"DL1"(open"OA")
seq"L2:x1(U):R5:L5:x1(0):R2:x2(0):R1:x2(U)"

StringFigures.TwistPassage — Type

A TwistPassage represents the twisting of one loop a certain number of times

julia> pass"<<<1"
pass"<<<1"

See also pass"", twist

StringFigures.ExtendPassage — Type

An ExtendPassage represents the extension of the string in order to make it taut. It corresponds with simplifications of the knot diagram which topologically don't affect the loop. It has no arguments. Represented in Storer with the symbol "|".

julia> pass"|"(seq"L1:x1(0):x1(U):R1")
seq"L1:R1"

See also simplify

StringFigures.PickPassage — Type

A PickPassage represents the action of picking a string with a given functor. Its arguments are:

fun::FrameRef : the functor (i.e. the picking finger)

away::Bool : on an "other hand" pick, if the first movement is away the executer (default)

arg::FrameRef : the argument (i.e. the finger holding the section of string being picked)

near::Bool : is it the near portion of the string?

over::Bool : does the finger travels above all other string in order to reach it?

above::Bool : does it pick the string from above?

In Storer, it is represented by F(A) in which Fis the functor, decorated with an arrow on top if the finger travels over other strings (theoverflag) and a downward pointing arrow "↓" if it picks the argument from above (theaboveflag). The argumentArepresents a framenode, appended with the letternorfif the string picked is onenear` the executer or not.

See also pick

StringFigures.MultiPickPassage — Type

A MultiPickPassage represents the action of picking a string with a given functor. Its arguments are:

pass::Vector{[PickPassage](@ref)} : A sequence indicating a 3D move in which the functor can pass either

over or under specific string segments attached to FrameNode.

julia> pass"L1o(L2n):L1u(L2f)"
pass"L1o(L2n):L1u(L2f)"

StringFigures.NavahoPassage — Type

A NavahoPassage represents the release of the lower loop in a two-loop finger. It is denoted by the "N" symbol in Storer.

julia> seq"L1.1:L2:R1:L1.0" |> pass"NL1"
seq"L1:x1(U):x2(0):L2:R1:x3(0):x1(0):x2(U):x4(U):x4(0):x3(U)"

See also pass""

StringFigures.PowerPassage — Type

A PowerPassage is just the repetition of another passage or sequence of passages

julia> pass"[<1#>1]^5"
pass"[<1#>1#]^5"

See also pass""

StringFigures.FrameRef — Type

A FrameRef is a reference to one loop attached to a FrameNode. l,m,u denote respectively the lowest, middle or top string on it. If there are more than 3 strings, then the second, third, etc are refered to as m1, m2, ...

StringFigures.BiFrameRef — Type

A BiFrameRef is a bilateral reference to a loop attached to a FrameNode.

StringFigures.@pass_str — Macro

pass"xxx" creates a Passage from string "xxx"

StringFigures.@fref_str — Macro

fref"xxx" creates a FrameRef or BiFrameRef

Calculus

StringFigures.StringCalculus — Type

A StringCalculus describes an algorithm or procedure that can be applied to a string. It is represented as list of transformations (Passages). You can build a StringCalculus by using the special notation calc"<seq>" where <seq> is a list of Passages, possibly separated by #. A StringCalculus can be:

Used as functions on a LinearSequence, producing a new LinearSequence

Multiplied to other StringCalculuss or Passages (concatenating the instructions)

Elevated to some power (repeating the same statements)

See also: StringProcedure

StringFigures.StringProcedure — Type

A StringProcedureconsists in an initialLinearSequenceplus aStringCalculus` to be applied to it. It represents the full "movie" of the figure construction. It can be

indexed and iterated to access each intermediate step

ploted to show all steps

See also proc""

StringFigures.@proc_str — Macro

proc"xxx" creates a StringProcedure from string "xxx"

julia> proc"OA::D1"
proc"OA::D1#"

julia> proc"OA::D1" |> last
seq"L2:x1(U):R5:L5:x1(0):R2:x2(0):x2(U)"

StringFigures.@calc_str — Macro

calc"xxx" produces a Calculus from "xxx"

julia> calc"DL1#DR1"
calc"DL1 # DR1"

Grammar

Input of Nodes, Linear sequences, Calculus, and full Procedures is specified by a PEG using the PEG.jl library. The full grammar is shown below.

FrameNode, CrossNode

@rule int =  r"\d+"
@rule fnode = r"[LR]" & int & ("." & int)[0:1]
@rule xnode = "x" & int & "(" & r"[0U]" & ")"
@rule snode = fnode, xnode

LinearSequence

@rule snodec = snode & ":"
@rule linseq = (snodec[*] & snode)
@rule opening = r"[0-9A-Za-z]*"p

FrameRef and BiFrameRef and Functor

@rule fref = r"l|u|m[1-9]?|" & r"[RL]?" & r"[0-9]"
@rule ffun = r"[LR]?" & int

Passage

@rule passage = extend_p, twist_p, release_p, navaho_p, multi_pick_p, pick_p
@rule extend_p = "|" & r"!*"p
@rule pick_p = ffun & r"[ou]"p & r"a?"p & r"t?"p & r"\("p & fref & r"[fn]"p & ")"
@rule pick_pp = pick_p & r":"p
@rule multi_pick_p = pick_pp[1:end] & pick_p
@rule release_p = "D" & fref
@rule navaho_p = "N" & ffun
@rule twist_p = r"(>+)|(<+)"p & fref
@rule power_p = "[" & calculus & "]^" & int > (_,f,_,n)->PowerPassage(f,n)

Calculus

@rule passages = (passage & r"#?"p)
@rule calculus = r""p & passages[*]

StringProcedure

@rule parenseq = "(" & linseq & ")"
@rule procedure = ((parenseq,opening) & r"::"p & calculus)